+++ examples/gql_example_cli_github/lib/__generated__/serializers.gql.dart
@@ -1,7 +1,7 @@
 import 'package:built_collection/built_collection.dart';
 import 'package:built_value/serializer.dart';
 import 'package:built_value/standard_json_plugin.dart' show StandardJsonPlugin;
-import 'package:gql_code_builder/src/serializers/operation_serializer.dart'
+import 'package:gql_code_builder_serializers/gql_code_builder_serializers.dart'
     show OperationSerializer;
 import 'package:gql_example_cli_github/__generated__/add_star.data.gql.dart'
     show GAddStarData, GAddStarData_action, GAddStarData_action_starrable;